requirements.txt
flask
flask_restplus
PyYAML
flask-swagger-ui==3.20.9